refactor: make filter menu its own component
